Health bar bug fix
git-svn-id: https://wiki.csc.calpoly.edu/svn/AsteroidBlaster@655 342171fe-4ba3-422f-b357-09b6f24955fe

+++ Utility/GameState.cpp
@@ -228,7 +228,7 @@ void GameState::update(double timeDiff) {
    updateText();
    
    weaponReadyBar->setAmount(ship->getCurrentWeaponCoolDown());
-   healthBar->setAmount((float) (ship->health / ship->healthMax));
+   healthBar->setAmount(((float) ship->health / (float) ship->healthMax));
    cube->update(timeDiff);
    minimap->update(timeDiff);
    spring->update(timeDiff);
